+++ noodles-gff/src/io/writer/line/record/source.rs
@@ -2,11 +2,14 @@ use std::io::{self, Write};
 
 use bstr::BStr;
 
+use super::percent_encode;
+
 pub(super) fn write_source<W>(writer: &mut W, source: &BStr) -> io::Result<()>
 where
     W: Write,
 {
-    writer.write_all(source)
+    let src = percent_encode(source);
+    writer.write_all(src.as_bytes())
 }
 
 #[cfg(test)]
@@ -15,9 +18,18 @@ mod tests {
 
     #[test]
     fn test_write_source() -> io::Result<()> {
+        fn t(buf: &mut Vec<u8>, source: &BStr, expected: &[u8]) -> io::Result<()> {
+            buf.clear();
+            write_source(buf, source)?;
+            assert_eq!(buf, expected);
+            Ok(())
+        }
+
         let mut buf = Vec::new();
-        write_source(&mut buf, BStr::new("NDLS"))?;
-        assert_eq!(buf, b"NDLS");
+
+        t(&mut buf, BStr::new("NDLS"), b"NDLS")?;
+        t(&mut buf, BStr::new("NDLS%"), b"NDLS%25")?;
+
         Ok(())
     }
 }